# specs/035-protected-resource-subresources/contracts/protected-resources.openapi.yaml
...
openapi: 3.0.3
info:
  title: Third-Party Service Protected Resources (delta)
  version: 1.0.0
  description: >
    Manage a third-party OAuth2 service's protected resource set as first-class, individually addressable
    sub-resources (keyed by percent-encoded normalized URI), plus the optimistic-concurrency change to the
    full-service update. Authentication/authorization are identical to the existing admin
    service-management endpoints (FR-014).
paths:
  /api/services/{service-id}/protected-resources:
    parameters:
      - $ref: '#/components/parameters/ServiceId'

    get:
      summary: List a service's protected resources (US4)
      operationId: listProtectedResources
      tags: [Services]
      responses:
        '200':
          description: The service's normalized protected resource URIs.
          headers:
            ETag:
              description: Service optimistic-concurrency token (version). Use as `If-Match` on replacement PUTs.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceSet' }
        '404': { $ref: '#/components/responses/ServiceNotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      summary: Add one protected resource from the collection (US1)
      description: >
        Adds the body `resource_uri` after server-side validation and normalization. This is an alternate
        collection-oriented add operation; PUT by percent-encoded member URI remains the canonical
        idempotent member operation. A newly added URI returns 201. A URI already owned by this service
        returns 200 with the set unchanged (FR-008); a URI owned by another service is a conflict (FR-007).
      operationId: createProtectedResource
      tags: [Services]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProtectedResourceCreateRequest' }
            examples:
              add:
                value: { resource_uri: "https://api.example.com/v2" }
      responses:
        '201':
          description: Resource added.
          headers:
            ETag: { schema: { type: string }, description: New service version. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '200':
          description: The service already owned the normalized URI; set unchanged.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '400':
          description: Missing request body or malformed/relative/empty/whitespace `resource_uri`; rejected before any state change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/ServiceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/services/{service-id}/protected-resources/{resource}:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
      - $ref: '#/components/parameters/ResourcePathSegment'

    put:
      summary: Add one protected resource (US1)
      description: >
        Idempotent add. The resource URI is the final path segment, fully percent-encoded and decoded +
        normalized server-side; there is no request body. Adding a URI (or one that normalizes to a URI)
        the service already owns succeeds with the set unchanged (FR-008). Adding a URI owned by another
        service is a conflict (FR-007).
      operationId: addProtectedResource
      tags: [Services]
      responses:
        '201':
          description: Resource added.
          headers:
            ETag: { schema: { type: string }, description: New service version. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '200':
          description: Idempotent — the service already owned the (normalized) URI; set unchanged.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '400': { $ref: '#/components/responses/InvalidResourceUri' }
        '404': { $ref: '#/components/responses/ServiceNotFound' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    patch:
      summary: Rename a protected resource in place (US3)
      description: >
        Changes the resource identified by the path segment (`{resource}`, the percent-encoded source URI)
        to the target URI in the body, atomically (no delete-then-add round trip). `resource == to` succeeds
        as a no-op. A `to` already owned by any service (including this one) is a conflict (FR-007). A
        `{resource}` not present on the service is not-found (FR-009).
      operationId: renameProtectedResource
      tags: [Services]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProtectedResourceRenameTarget' }
            examples:
              rename:
                value: { to: "https://api.example.com/v2" }
      responses:
        '200':
          description: Resource renamed (or no-op when source == to).
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '400': { $ref: '#/components/responses/InvalidResourceUri' }
        '404':
          description: Service not found, or `{resource}` is not a resource of this service (FR-009/FR-010).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/ResourceConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      summary: Remove one protected resource (US2)
      description: >
        Removes the resource identified by the path segment (`{resource}`, the percent-encoded URI, decoded
        and normalized server-side before matching). Removing a URI the service does not own is not-found
        (FR-009). Immediately stops new token-exchange resolution for that URI (FR-013).
      operationId: removeProtectedResource
      tags: [Services]
      responses:
        '200':
          description: Resource removed. Returns the removed normalized URI and the resulting set (FR-016).
          headers:
            ETag:
              description: New strong service ETag.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProtectedResourceMutationResult' }
        '400': { $ref: '#/components/responses/InvalidResourceUri' }
        '404':
          description: Service not found, or the URI is not a resource of this service (FR-009/FR-010).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ---- Behavioral delta on the existing endpoint (documented here; merged into the canonical path) ----
...
components:
  parameters:
    ServiceId:
      name: service-id
      in: path
      required: true
      schema: { type: string, format: uuid }
    ResourcePathSegment:
      name: resource
      in: path
      required: true
      description: >
        The target protected resource's normalized URI, fully RFC 3986 percent-encoded as a SINGLE path
        segment (`:`→`%3A`, `/`→`%2F`, `?`→`%3F`, `#`→`%23`). Decoded (`url.PathUnescape`) and normalized
        (trailing-slash trim) server-side before matching. A segment that does not decode to a well-formed
        absolute URI (scheme + host) is rejected with 400. Note: an upstream proxy/gateway that rejects or
        rewrites encoded slashes must be configured to pass `%2F` through (deployment consideration).
      schema: { type: string }
      example: "https%3A%2F%2Fapi.example.com%2Fv2"
  schemas:
    ProtectedResourceSet:
      type: object
      required: [protected_resources]
      properties:
        protected_resources:
          type: array
          items: { type: string, format: uri }
          example: ["https://api.github.com", "https://github.com/api/v3"]
    ProtectedResourceCreateRequest:
      type: object
      required: [resource_uri]
      properties:
        resource_uri:
          type: string
          format: uri
          description: Absolute resource URI (scheme + host). Normalized (trailing-slash trim) server-side.
          example: "https://api.example.com/v2"
    ProtectedResourceRenameTarget:
      type: object
      required: [to]
      properties:
        to:
          type: string
          format: uri
          description: Absolute target URI (scheme + host). Normalized (trailing-slash trim) server-side.
          example: "https://api.example.com/v2"
    ProtectedResourceMutationResult:
      description: Enough resulting state to confirm the outcome without a full-service fetch (FR-016).
      type: object
      required: [resource, protected_resources]
      properties:
        resource:
          type: string
          format: uri
          description: The normalized URI affected by the operation.
        protected_resources:
          type: array
          items: { type: string, format: uri }
          description: The service's resulting resource set.
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
  responses:
    ServiceNotFound:
      description: No service with the given id (FR-010).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    InvalidResourceUri:
      description: Malformed/relative/empty/whitespace URI — rejected before any state change (FR-006).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    ResourceConflict:
      description: URI already owned by another service, or rename target already present (FR-007).
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json: { schema: { $ref: '#/components/schemas/Error' } }